+++ R/sparklyr-spark-connect.R
@@ -63,32 +63,19 @@ spark_connect_method.spark_method_databricks_connect <- function(
     ...) {
   args <- list(...)
   cluster_id <- args$cluster_id
+  serverless <- args$serverless %||% FALSE
+  profile <- args$profile %||% NULL
   token <- args$token
   envname <- args$envname
   host_sanitize <- args$host_sanitize %||% TRUE
   silent <- args$silent %||% FALSE
 
   method <- method[[1]]
+
   token <- databricks_token(token, fail = FALSE)
   cluster_id <- cluster_id %||% Sys.getenv("DATABRICKS_CLUSTER_ID")
-  master <- databricks_host(master, fail = FALSE)
-  if (host_sanitize && master != "") {
-    master <- sanitize_host(master, silent)
-  }
 
-  cluster_info <- NULL
-  if (cluster_id != "" && master != "" && token != "") {
-    cluster_info <- databricks_dbr_version_name(
-      cluster_id = cluster_id,
-      host = master,
-      token = token,
-      silent = silent
-    )
-    if (is.null(version)) {
-      version <- cluster_info$version
-    }
-  }
-
+  # load python env
   envname <- use_envname(
     backend = "databricks",
     version = version,
@@ -102,16 +89,71 @@ spark_connect_method.spark_method_databricks_connect <- function(
     return(invisible)
   }
 
-  db <- import_check("databricks.connect", envname, silent)
+  # load python libs
+  dbc <- import_check("databricks.connect", envname, silent)
+  db_sdk <- import_check("databricks.sdk", envname, silent = TRUE)
+
+  # SDK behaviour
+  # https://databricks-sdk-py.readthedocs.io/en/latest/authentication.html#default-authentication-flow
+
+  conf_args <- list()
+
+  # the profile as specified - which has a default of 'DEFAULT'
+  # otherwise, if a token is found, propagate to SDK config
+
+  # TODO: emit messages about connection here?
+  # specific vars taken priority, profile only works when no env vars are set
+  if (token != "" && master != "") {
+    conf_args$host <-  master
+    conf_args$token <- token
+    conf_args$auth_type <- "pat"
+    databricks_desktop_login(host = master)
+  } else if (!is.null(profile)) {
+    conf_args$profile <- profile
+    databricks_desktop_login(profile = profile)
+  }
+
+  # serverless config related settings
+  if (serverless) {
+    conf_args$serverless_compute_id <- "auto"
+  } else {
+    conf_args$cluster_id <- cluster_id
+  }
+
+  sdk_config <- db_sdk$core$Config(!!!conf_args)
+
+  # create workspace client
+  sdk_client <- db_sdk$WorkspaceClient(config = sdk_config)
+
+  # if serverless is TRUE, cluster_id is overruled (set to NULL)
+  cluster_info <- NULL
+  if (!serverless) {
+    if (cluster_id != "" && master != "" && token != "") {
+      cluster_info <- databricks_dbr_version_name(
+        cluster_id = cluster_id,
+        client = sdk_client,
+        silent = silent
+      )
+      if (is.null(version)) {
+        version <- cluster_info$version
+      }
+    }
+  } else {
+    cluster_id <- NULL
+  }
 
   if (!is.null(cluster_info)) {
     msg <- "{.header Connecting to} {.emph '{cluster_info$name}'}"
     msg_done <- "{.header Connected to:} {.emph '{cluster_info$name}'}"
     master_label <- glue("{cluster_info$name} ({cluster_id})")
-  } else {
+  } else if (!serverless) {
     msg <- "{.header Connecting to} {.emph '{cluster_id}'}"
     msg_done <- "{.header Connected to:} '{.emph '{cluster_id}'}'"
     master_label <- glue("Databricks Connect - Cluster: {cluster_id}")
+  } else if (serverless) {
+    msg <- "{.header Connecting to} {.emph serverless}"
+    msg_done <- "{.header Connected to:} '{.emph serverless}'"
+    master_label <- glue("Databricks Connect - Cluster: serverless")
   }
 
   if (!silent) {
@@ -119,17 +161,8 @@ spark_connect_method.spark_method_databricks_connect <- function(
     cli_progress_step(msg, msg_done)
   }
 
-  remote_args <- list()
-  if (master != "") remote_args$host <- master
-  if (token != "") remote_args$token <- token
-  if (cluster_id != "") remote_args$cluster_id <- cluster_id
-
-  databricks_session <- function(...) {
-    user_agent <- build_user_agent()
-    db$DatabricksSession$builder$remote(...)$userAgent(user_agent)
-  }
-
-  conn <- exec(databricks_session, !!!remote_args)
+  user_agent <- build_user_agent()
+  conn <- dbc$DatabricksSession$builder$sdkConfig(sdk_config)$userAgent(user_agent)
 
   if (!silent) {
     cli_progress_done()
@@ -141,6 +174,7 @@ spark_connect_method.spark_method_databricks_connect <- function(
     master_label = master_label,
     con_class = "connect_databricks",
     cluster_id = cluster_id,
+    serverless = serverless,
     method = method,
     config = config
   )
@@ -151,6 +185,7 @@ initialize_connection <- function(
     master_label,
     con_class,
     cluster_id = NULL,
+    serverless = NULL,
     method = NULL,
     config = NULL) {
   warnings <- import("warnings")
@@ -173,12 +208,15 @@ initialize_connection <- function(
     "ignore",
     message = "Index.format is deprecated and will be removed in a future version"
   )
+
   session <- conn$getOrCreate()
   get_version <- try(session$version, silent = TRUE)
   if (inherits(get_version, "try-error")) databricks_dbr_error(get_version)
-  session$conf$set("spark.sql.session.localRelationCacheThreshold", 1048576L)
-  session$conf$set("spark.sql.execution.arrow.pyspark.enabled", "true")
-  session$conf$set("spark.sql.execution.arrow.pyspark.fallback.enabled", "false")
+  if (!serverless) {
+    session$conf$set("spark.sql.session.localRelationCacheThreshold", 1048576L)
+    session$conf$set("spark.sql.execution.arrow.pyspark.enabled", "true")
+    session$conf$set("spark.sql.execution.arrow.pyspark.fallback.enabled", "false")
+  }
 
   # do we need this `spark_context` object?
   spark_context <- list(spark_context = session)